+++ deepcave/cli.py
@@ -1,5 +1,3 @@
-from typing import Any, List
-
 import multiprocessing
 import subprocess
 from pathlib import Path
@@ -49,4 +47,7 @@ def execute(_) -> None:
 
 
 def main() -> None:
-    app.run(execute)
+    try:
+        app.run(execute)
+    except KeyboardInterrupt:
+        exit("KeyboardInterrupt.")