+++ deepcave/evaluators/epm/random_forest.py
@@ -180,7 +180,7 @@ class RandomForest:
 
     def _check_dimensions(self, X: np.ndarray, Y: Optional[np.ndarray] = None) -> None:
         """
-        Checks if the dimensions of X and Y are correct wrt features.
+        Checks if the dimensions of X and Y are correct with respect to features.
 
         Parameters
         ----------