+++ deepcave/evaluators/fanova.py
@@ -33,7 +33,7 @@ class fANOVA:
         seed: int = 0,
     ) -> None:
         """
-        Get the data wrt budget and trains the forest on the encoded data.
+        Get the data with respect to budget and trains the forest on the encoded data.
 
         Note
         ----